CPS and better Either handling for jsonRequest.
I did a lot of development on another branch, and concluded that the
overall code would flow more smoothly if I adopted a CPS style for
jsonRequest, accepting a function that would transform the raw results
into what the user was actually looking for.

The handling of the Either is a little better, too, I think.