Rewrite internal tests to use hunit and lens, not hspec.
While hspec might be the better option long-term, the version we have in
Debian right now is too old to do what I need.

Moving to lens is just for my own sanity.  Of course, because of my dev
environment, and cabal's limitations, I have to stick with lens < 4.4
for the moment.