Refactor low-level request code
It turns out that _config support required that we be able to parse any
JSON values, not just structures.  Although I could have just dropped in
the 'value' parser rather than the 'json' parser, I decided to refactor
so that we would only use the more liberal parser where we needed to.

Eventually we will need another round of refactoring, as we get rid of
the assumptions of JSON that are still embedded there, but that's a task
for another day.