Fix initial doc queries.
First off, they weren't handling the `Maybe DocRev` correctly.  And then
once that was solved, they needed to be able to handle 304 return
values, which needed a change in `checkStatusCode`, as well as checks to
return an appropriate value (in this case, `Null`, since that's
otherwise *not* a legit document).